worker/ingestion/requirements.txt
# Ingestion Worker Dependencies

# --- Core ---
celery[redis]>=5.3
celery-pool-asyncio
psycopg2-binary>=2.9
pydantic>=2.0
pydantic-settings>=2.0

# --- Database ---
sqlalchemy>=2.0
asyncpg>=0.25.0 # For async PostgreSQL access
psycopg2-binary # Sometimes needed by Alembic even if using asyncpg

# --- Git & Data  ---
GitPython>=3.1.30
numpy>1.20.0,<1.27.0
pandas>=1.5.0

# --- GitHub API Access ---
requests>=2.20.0
python-dateutil>=2.8.0
